Moved solidgauge series to subfolder
