+++ js/Series/SolidGauge/SolidGaugeSeries.js
@@ -9,82 +9,39 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
-
 'use strict';
-
-import type ColorAxis from '../Core/Axis/ColorAxis';
-import type ColorType from '../Core/Color/ColorType';
-import type GradientColor from '../Core/Color/GradientColor';
-import type GaugePoint from './Gauge/GaugePoint';
-import type GaugePointOptions from './Gauge/GaugePointOptions';
-import type GaugeSeriesOptions from './Gauge/GaugeSeriesOptions';
-import type RadialAxis from '../Core/Axis/RadialAxis';
-import type { SeriesStatesOptions } from '../Core/Series/SeriesOptions';
-import type SVGAttributes from '../Core/Renderer/SVG/SVGAttributes';
-import type SVGPath from '../Core/Renderer/SVG/SVGPath';
-import BaseSeries from '../Core/Series/Series.js';
-import GaugeSeries from './Gauge/GaugeSeries.js';
-import Color from '../Core/Color/Color.js';
-const {
-    parse: color
-} = Color;
-import H from '../Core/Globals.js';
-const {
-    Renderer
-} = H;
-import LegendSymbolMixin from '../Mixins/LegendSymbol.js';
-import U from '../Core/Utilities.js';
-const {
-    clamp,
-    extend,
-    isNumber,
-    merge,
-    pick,
-    pInt,
-    wrap
-} = U;
-
-/**
- * Internal types
- * @private
- */
-declare global {
-    namespace Highcharts {
-        class SolidGaugePoint extends GaugePoint {
-            options: SolidGaugePointOptions;
-            series: SolidGaugeSeries;
-        }
-        interface SolidGaugePointOptions extends GaugePointOptions {
-            innerRadius?: (number|string);
-            radius?: (number|string);
-        }
-        interface SolidGaugeSeriesOptions extends GaugeSeriesOptions {
-            innerRadius?: (number|string);
-            linecap?: string;
-            overshoot?: number;
-            radius?: (number|string);
-            rounded?: boolean;
-            states?: SeriesStatesOptions<SolidGaugeSeries>;
-            threshold?: number;
-        }
-        interface SymbolOptionsObject {
-            rounded?: boolean;
-        }
-    }
-}
-
-import '../Core/Options.js';
-
+var __extends = (this && this.__extends) || (function () {
+    var extendStatics = function (d, b) {
+        extendStatics = Object.setPrototypeOf ||
+            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+        return extendStatics(d, b);
+    };
+    return function (d, b) {
+        extendStatics(d, b);
+        function __() { this.constructor = d; }
+        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+    };
+})();
+import BaseSeries from '../../Core/Series/Series.js';
+import GaugeSeries from '../Gauge/GaugeSeries.js';
+import Color from '../../Core/Color/Color.js';
+var color = Color.parse;
+import H from '../../Core/Globals.js';
+var Renderer = H.Renderer;
+import LegendSymbolMixin from '../../Mixins/LegendSymbol.js';
+import U from '../../Core/Utilities.js';
+var clamp = U.clamp, extend = U.extend, isNumber = U.isNumber, merge = U.merge, pick = U.pick, pInt = U.pInt, wrap = U.wrap;
+import '../../Core/Options.js';
 /**
  * Additional options, depending on the actual symbol drawn.
  *
  * @interface Highcharts.SymbolOptionsObject
- *//**
- * Whether to draw rounded edges.
- * @name Highcharts.SymbolOptionsObject#rounded
- * @type {boolean|undefined}
- */
-
+ */ /**
+* Whether to draw rounded edges.
+* @name Highcharts.SymbolOptionsObject#rounded
+* @type {boolean|undefined}
+*/
 /**
  * Symbol definition of an arc with round edges.
  *
@@ -109,86 +66,34 @@ import '../Core/Options.js';
  * @return {Highcharts.SVGPathArray}
  *         Path of the created arc.
  */
-wrap(
-    Renderer.prototype.symbols,
-    'arc',
-    function (
-        proceed: Function,
-        x: number,
-        y: number,
-        w: number,
-        h: number,
-        options: Highcharts.SymbolOptionsObject
-    ): SVGPath {
-        var arc = proceed,
-            path: SVGPath = arc(x, y, w, h, options);
-
-        if (options.rounded) {
-            var r = options.r || w,
-                smallR = (r - (options.innerR || 0)) / 2,
-                outerArcStart = path[0],
-                innerArcStart = path[2];
-
-            if (outerArcStart[0] === 'M' && innerArcStart[0] === 'L') {
-                const x1 = outerArcStart[1],
-                    y1 = outerArcStart[2],
-                    x2 = innerArcStart[1],
-                    y2 = innerArcStart[2],
-                    roundStart: SVGPath.Arc = ['A', smallR, smallR, 0, 1, 1, x1, y1],
-                    roundEnd: SVGPath.Arc = ['A', smallR, smallR, 0, 1, 1, x2, y2];
-
-                // Replace the line segment and the last close segment
-                path[2] = roundEnd;
-                path[4] = roundStart;
-            }
+wrap(Renderer.prototype.symbols, 'arc', function (proceed, x, y, w, h, options) {
+    var arc = proceed, path = arc(x, y, w, h, options);
+    if (options.rounded) {
+        var r = options.r || w, smallR = (r - (options.innerR || 0)) / 2, outerArcStart = path[0], innerArcStart = path[2];
+        if (outerArcStart[0] === 'M' && innerArcStart[0] === 'L') {
+            var x1 = outerArcStart[1], y1 = outerArcStart[2], x2 = innerArcStart[1], y2 = innerArcStart[2], roundStart = ['A', smallR, smallR, 0, 1, 1, x1, y1], roundEnd = ['A', smallR, smallR, 0, 1, 1, x2, y2];
+            // Replace the line segment and the last close segment
+            path[2] = roundEnd;
+            path[4] = roundStart;
         }
-
-        return path;
     }
-);
-
-/* eslint-disable valid-jsdoc */
-
+    return path;
+});
 /**
  * @private
  */
-interface SolidGaugeAxis extends RadialAxis {
-    dataClasses: ColorAxis['dataClasses'];
-    options: SolidGaugeAxis.Options;
-    stops: ColorAxis['stops'];
-    initDataClasses(userOptions: ColorAxis.Options): void;
-    initStops(userOptions: ColorAxis.Options): void;
-    toColor(
-        value: number,
-        point: Highcharts.SolidGaugePoint
-    ): (ColorType|undefined);
-}
-
-/**
- * @private
- */
-namespace SolidGaugeAxis {
-
+var SolidGaugeAxis;
+(function (SolidGaugeAxis) {
     /* *
      *
      *  Interfaces
      *
      * */
-
-    export interface Options extends ColorAxis.Options {
-        dataClassColor?: ColorAxis.Options['dataClassColor'];
-        dataClasses?: ColorAxis.Options['dataClasses'];
-        maxColor?: ColorAxis.Options['maxColor'];
-        minColor?: ColorAxis.Options['minColor'];
-        stops?: ColorAxis.Options['stops'];
-    }
-
     /* *
      *
      *  Constants
      *
      * */
-
     /**
      * These methods are defined in the ColorAxis object, and copied here.
      * @private
@@ -196,84 +101,49 @@ namespace SolidGaugeAxis {
      * @todo
      * If we implement an AMD system we should make ColorAxis a dependency.
      */
-    const methods = {
-
-        initDataClasses: function (
-            this: SolidGaugeAxis,
-            userOptions: ColorAxis.Options
-        ): void {
-            var chart = this.chart,
-                dataClasses: Array<ColorAxis.DataClassesOptions>,
-                colorCounter = 0,
-                options = this.options;
-
+    var methods = {
+        initDataClasses: function (userOptions) {
+            var chart = this.chart, dataClasses, colorCounter = 0, options = this.options;
             this.dataClasses = dataClasses = [];
-
-            (userOptions.dataClasses as any).forEach(function (
-                dataClass: ColorAxis.DataClassesOptions,
-                i: number
-            ): void {
-                var colors: (Array<string>|undefined);
-
+            userOptions.dataClasses.forEach(function (dataClass, i) {
+                var colors;
                 dataClass = merge(dataClass);
                 dataClasses.push(dataClass);
                 if (!dataClass.color) {
                     if (options.dataClassColor === 'category') {
                         colors = chart.options.colors;
-                        dataClass.color = (colors as any)[colorCounter++];
+                        dataClass.color = colors[colorCounter++];
                         // loop back to zero
-                        if (colorCounter === (colors as any).length) {
+                        if (colorCounter === colors.length) {
                             colorCounter = 0;
                         }
-                    } else {
-                        dataClass.color = color(options.minColor).tweenTo(
-                            color(options.maxColor),
-                            i / ((userOptions.dataClasses as any).length - 1)
-                        );
+                    }
+                    else {
+                        dataClass.color = color(options.minColor).tweenTo(color(options.maxColor), i / (userOptions.dataClasses.length - 1));
                     }
                 }
             });
         },
-
-        initStops: function (
-            this: SolidGaugeAxis,
-            userOptions: ColorAxis.Options
-        ): void {
+        initStops: function (userOptions) {
             this.stops = userOptions.stops || [
-                [0, this.options.minColor as any],
-                [1, this.options.maxColor as any]
+                [0, this.options.minColor],
+                [1, this.options.maxColor]
             ];
-            this.stops.forEach(function (
-                stop: GradientColor['stops'][0]
-            ): void {
+            this.stops.forEach(function (stop) {
                 stop.color = color(stop[1]);
             });
         },
         // Translate from a value to a color
-        toColor: function (
-            this: SolidGaugeAxis,
-            value: number,
-            point: Highcharts.SolidGaugePoint
-        ): (ColorType|undefined) {
-            var pos: number,
-                stops = this.stops,
-                from: (number|GradientColor['stops'][0]|undefined),
-                to: (number|GradientColor['stops'][0]|undefined),
-                color: (ColorType|undefined),
-                dataClasses = this.dataClasses,
-                dataClass: (ColorAxis.DataClassesOptions|undefined),
-                i: (number|undefined);
-
+        toColor: function (value, point) {
+            var pos, stops = this.stops, from, to, color, dataClasses = this.dataClasses, dataClass, i;
             if (dataClasses) {
                 i = dataClasses.length;
                 while (i--) {
                     dataClass = dataClasses[i];
                     from = dataClass.from;
                     to = dataClass.to;
-                    if (
-                        (typeof from === 'undefined' || value >= from) &&
-                        (typeof to === 'undefined' || value <= to)
-                    ) {
+                    if ((typeof from === 'undefined' || value >= from) &&
+                        (typeof to === 'undefined' || value <= to)) {
                         color = dataClass.color;
                         if (point) {
                             point.dataClass = i;
@@ -281,9 +151,8 @@ namespace SolidGaugeAxis {
                         break;
                     }
                 }
-
-            } else {
-
+            }
+            else {
                 if (this.logarithmic) {
                     value = this.val2lin(value);
                 }
@@ -296,35 +165,27 @@ namespace SolidGaugeAxis {
                 }
                 from = stops[i] || stops[i + 1];
                 to = stops[i + 1] || from;
-
                 // The position within the gradient
-                pos = (1 - ((to as any)[0] - pos) / (((to as any)[0] -
-                    (from as any)[0]) || 1));
-
-                color = (from as any).color.tweenTo(
-                    (to as any).color,
-                    pos
-                );
+                pos = (1 - (to[0] - pos) / ((to[0] -
+                    from[0]) || 1));
+                color = from.color.tweenTo(to.color, pos);
             }
             return color;
         }
     };
-
     /* *
      *
      *  Functions
      *
      * */
-
     /**
      * @private
      */
-    export function init(axis: RadialAxis): void {
+    function init(axis) {
         extend(axis, methods);
     }
-
-}
-
+    SolidGaugeAxis.init = init;
+})(SolidGaugeAxis || (SolidGaugeAxis = {}));
 /**
  * A solid gauge is a circular gauge where the value is indicated by a filled
  * arc, and the color of the arc may variate with the value.
@@ -338,7 +199,7 @@ namespace SolidGaugeAxis {
  * @requires     modules/solid-gauge
  * @optionparent plotOptions.solidgauge
  */
-var solidGaugeOptions: Highcharts.SolidGaugeSeriesOptions = {
+var solidGaugeOptions = {
     /**
      * The inner radius for points in a solid gauge. Can be given as a number
      * (pixels) or percentage string.
@@ -352,7 +213,6 @@ var solidGaugeOptions: Highcharts.SolidGaugeSeriesOptions = {
      * @product   highcharts
      * @apioption plotOptions.solidgauge.innerRadius
      */
-
     /**
      * Whether the strokes of the solid gauge should be `round` or `square`.
      *
@@ -366,7 +226,6 @@ var solidGaugeOptions: Highcharts.SolidGaugeSeriesOptions = {
      * @validvalue ["square", "round"]
      * @apioption  plotOptions.solidgauge.linecap
      */
-
     /**
      * Allow the gauge to overshoot the end of the perimeter axis by this
      * many degrees. Say if the gauge axis goes from 0 to 60, a value of
@@ -379,7 +238,6 @@ var solidGaugeOptions: Highcharts.SolidGaugeSeriesOptions = {
      * @product   highcharts
      * @apioption plotOptions.solidgauge.overshoot
      */
-
     /**
      * The outer radius for points in a solid gauge. Can be given as a number
      * (pixels) or percentage string.
@@ -393,7 +251,6 @@ var solidGaugeOptions: Highcharts.SolidGaugeSeriesOptions = {
      * @product   highcharts
      * @apioption plotOptions.solidgauge.radius
      */
-
     /**
      * Wether to draw rounded edges on the gauge.
      *
@@ -406,7 +263,6 @@ var solidGaugeOptions: Highcharts.SolidGaugeSeriesOptions = {
      * @product   highcharts
      * @apioption plotOptions.solidgauge.rounded
      */
-
     /**
      * The threshold or base level for the gauge.
      *
@@ -418,173 +274,92 @@ var solidGaugeOptions: Highcharts.SolidGaugeSeriesOptions = {
      * @product   highcharts
      * @apioption plotOptions.solidgauge.threshold
      */
-
     /**
      * Whether to give each point an individual color.
      */
     colorByPoint: true,
-
     dataLabels: {
         y: 0
     }
-
 };
-
-
 /* *
  *
  *  Class
  *
  * */
-
-class SolidGaugeSeries extends GaugeSeries {
-
-    /* *
-     *
-     *  Static properties
-     *
-     * */
-
-    public static defaultOptions: Highcharts.SolidGaugeSeriesOptions = merge(GaugeSeries.defaultOptions,
-        solidGaugeOptions as Highcharts.SolidGaugeSeriesOptions);
-
-    /* *
-     *
-     *  Properties
-     *
-     * */
-
-    public data: Array<Highcharts.SolidGaugePoint> = void 0 as any;
-    public points: Array<Highcharts.SolidGaugePoint> = void 0 as any;
-    public options: Highcharts.SolidGaugeSeriesOptions = void 0 as any;
-
-    public axis: SolidGaugeAxis = void 0 as any;
-    public yAxis: SolidGaugeAxis = void 0 as any;
-    public startAngleRad: SolidGaugeSeries['thresholdAngleRad'] = void 0 as any;
-    public thresholdAngleRad: number = void 0 as any;
+var SolidGaugeSeries = /** @class */ (function (_super) {
+    __extends(SolidGaugeSeries, _super);
+    function SolidGaugeSeries() {
+        /* *
+         *
+         *  Static properties
+         *
+         * */
+        var _this = _super !== null && _super.apply(this, arguments) || this;
+        /* *
+         *
+         *  Properties
+         *
+         * */
+        _this.data = void 0;
+        _this.points = void 0;
+        _this.options = void 0;
+        _this.axis = void 0;
+        _this.yAxis = void 0;
+        _this.startAngleRad = void 0;
+        _this.thresholdAngleRad = void 0;
+        return _this;
+    }
     /* *
      *
      *  Functions
      *
      * */
-
     // Extend the translate function to extend the Y axis with the necessary
     // decoration (#5895).
-    public translate(): void {
+    SolidGaugeSeries.prototype.translate = function () {
         var axis = this.yAxis;
-
         SolidGaugeAxis.init(axis);
-
         // Prepare data classes
         if (!axis.dataClasses && axis.options.dataClasses) {
             axis.initDataClasses(axis.options);
         }
         axis.initStops(axis.options);
-
         // Generate points and inherit data label position
         H.seriesTypes.gauge.prototype.translate.call(this);
-    }
-
+    };
     // Draw the points where each point is one needle.
-    public drawPoints(): void {
-        var series = this,
-            yAxis = series.yAxis,
-            center = yAxis.center,
-            options = series.options,
-            renderer = series.chart.renderer,
-            overshoot = options.overshoot,
-            overshootVal = isNumber(overshoot) ?
-                overshoot / 180 * Math.PI :
-                0,
-            thresholdAngleRad: (number | undefined);
-
+    SolidGaugeSeries.prototype.drawPoints = function () {
+        var series = this, yAxis = series.yAxis, center = yAxis.center, options = series.options, renderer = series.chart.renderer, overshoot = options.overshoot, overshootVal = isNumber(overshoot) ?
+            overshoot / 180 * Math.PI :
+            0, thresholdAngleRad;
         // Handle the threshold option
         if (isNumber(options.threshold)) {
-            thresholdAngleRad = yAxis.startAngleRad + (yAxis.translate(
-                options.threshold,
-                null,
-                null,
-                null,
-                true
-            ) as any);
+            thresholdAngleRad = yAxis.startAngleRad + yAxis.translate(options.threshold, null, null, null, true);
         }
-        this.thresholdAngleRad = pick(
-            thresholdAngleRad, yAxis.startAngleRad
-        );
-
-
-        series.points.forEach(function (
-            point: Highcharts.SolidGaugePoint
-        ): void {
+        this.thresholdAngleRad = pick(thresholdAngleRad, yAxis.startAngleRad);
+        series.points.forEach(function (point) {
             // #10630 null point should not be draw
             if (!point.isNull) { // condition like in pie chart
-                var graphic = point.graphic,
-                    rotation = (yAxis.startAngleRad +
-                        (yAxis.translate(
-                            point.y as any,
-                            null,
-                            null,
-                            null,
-                            true
-                        ) as any)),
-                    radius = ((
-                        pInt(
-                            pick(
-                                point.options.radius,
-                                options.radius,
-                                100
-                            )
-                        ) * center[2]
-                    ) / 200),
-                    innerRadius = ((
-                        pInt(
-                            pick(
-                                point.options.innerRadius,
-                                options.innerRadius,
-                                60
-                            )
-                        ) * center[2]
-                    ) / 200),
-                    shapeArgs: (SVGAttributes | undefined),
-                    d: (string | SVGPath | undefined),
-                    toColor = yAxis.toColor(point.y as any, point),
-                    axisMinAngle = Math.min(
-                        yAxis.startAngleRad,
-                        yAxis.endAngleRad
-                    ),
-                    axisMaxAngle = Math.max(
-                        yAxis.startAngleRad,
-                        yAxis.endAngleRad
-                    ),
-                    minAngle,
-                    maxAngle;
-
+                var graphic = point.graphic, rotation = (yAxis.startAngleRad +
+                    yAxis.translate(point.y, null, null, null, true)), radius = ((pInt(pick(point.options.radius, options.radius, 100)) * center[2]) / 200), innerRadius = ((pInt(pick(point.options.innerRadius, options.innerRadius, 60)) * center[2]) / 200), shapeArgs, d, toColor = yAxis.toColor(point.y, point), axisMinAngle = Math.min(yAxis.startAngleRad, yAxis.endAngleRad), axisMaxAngle = Math.max(yAxis.startAngleRad, yAxis.endAngleRad), minAngle, maxAngle;
                 if (toColor === 'none') { // #3708
                     toColor = point.color || series.color || 'none';
                 }
                 if (toColor !== 'none') {
                     point.color = toColor;
                 }
-
                 // Handle overshoot and clipping to axis max/min
-                rotation = clamp(
-                    rotation,
-                    axisMinAngle - overshootVal,
-                    axisMaxAngle + overshootVal
-                );
-
+                rotation = clamp(rotation, axisMinAngle - overshootVal, axisMaxAngle + overshootVal);
                 // Handle the wrap option
                 if (options.wrap === false) {
                     rotation = clamp(rotation, axisMinAngle, axisMaxAngle);
                 }
-
                 minAngle = Math.min(rotation, series.thresholdAngleRad);
                 maxAngle = Math.max(rotation, series.thresholdAngleRad);
-
                 if (maxAngle - minAngle > 2 * Math.PI) {
                     maxAngle = minAngle + 2 * Math.PI;
                 }
-
                 point.shapeArgs = shapeArgs = {
                     x: center[0],
                     y: center[1],
@@ -595,22 +370,21 @@ class SolidGaugeSeries extends GaugeSeries {
                     rounded: options.rounded
                 };
                 point.startR = radius; // For PieSeries.animate
-
                 if (graphic) {
                     d = shapeArgs.d;
                     graphic.animate(extend({ fill: toColor }, shapeArgs));
                     if (d) {
                         shapeArgs.d = d; // animate alters it
                     }
-                } else {
+                }
+                else {
                     point.graphic = graphic = renderer.arc(shapeArgs)
                         .attr({
-                            fill: toColor,
-                            'sweep-flag': 0
-                        })
+                        fill: toColor,
+                        'sweep-flag': 0
+                    })
                         .add(series.group);
                 }
-
                 if (!series.chart.styledMode) {
                     if (options.linecap !== 'square') {
                         graphic.attr({
@@ -623,64 +397,32 @@ class SolidGaugeSeries extends GaugeSeries {
                         'stroke-width': options.borderWidth || 0
                     });
                 }
-
                 if (graphic) {
                     graphic.addClass(point.getClassName(), true);
                 }
             }
         });
-    }
-
+    };
     // Extend the pie slice animation by animating from start angle and up.
-    animate(init?: boolean): void {
+    SolidGaugeSeries.prototype.animate = function (init) {
         if (!init) {
             this.startAngleRad = this.thresholdAngleRad;
             H.seriesTypes.pie.prototype.animate.call(this, init);
         }
-    }
-}
-
-/* *
- *
- *  Prototype properties
- *
- * */
-
-interface SolidGaugeSeries extends GaugeSeries {
-    pointClass: typeof Highcharts.SolidGaugePoint;
-    drawLegendSymbol: typeof LegendSymbolMixin.drawRectangle;
-}
+    };
+    SolidGaugeSeries.defaultOptions = merge(GaugeSeries.defaultOptions, solidGaugeOptions);
+    return SolidGaugeSeries;
+}(GaugeSeries));
 extend(SolidGaugeSeries.prototype, {
     drawLegendSymbol: LegendSymbolMixin.drawRectangle
 });
-
-
-/* *
- *
- *  Registry
- *
- * */
-
-
-/**
- * @private
- */
-declare module '../Core/Series/SeriesType' {
-    interface SeriesTypeRegistry {
-        solidgauge: typeof SolidGaugeSeries;
-    }
-}
-
 BaseSeries.registerSeriesType('solidgauge', SolidGaugeSeries);
-
 /* *
  *
  *  Default export
  *
  * */
-
 export default SolidGaugeSeries;
-
 /**
  * A `solidgauge` series. If the [type](#series.solidgauge.type) option is not
  * specified, it is inherited from [chart.type](#chart.type).
@@ -697,7 +439,6 @@ export default SolidGaugeSeries;
  * @requires  modules/solid-gauge
  * @apioption series.solidgauge
  */
-
 /**
  * An array of data points for the series. For the `solidgauge` series
  * type, points can be given in the following ways:
@@ -737,7 +478,6 @@ export default SolidGaugeSeries;
  * @product   highcharts
  * @apioption series.solidgauge.data
  */
-
 /**
  * The inner radius of an individual point in a solid gauge. Can be given as a
  * number (pixels) or percentage string.
@@ -750,7 +490,6 @@ export default SolidGaugeSeries;
  * @product   highcharts
  * @apioption series.solidgauge.data.innerRadius
  */
-
 /**
  * The outer radius of an individual point in a solid gauge. Can be
  * given as a number (pixels) or percentage string.
@@ -763,5 +502,4 @@ export default SolidGaugeSeries;
  * @product   highcharts
  * @apioption series.solidgauge.data.radius
  */
-
 ''; // adds doclets above to transpiled file
