+++ js/masters/modules/solid-gauge.src.js
@@ -11,4 +11,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Series/SolidGaugeSeries.js';
+import '../../Series/SolidGauge/SolidGaugeSeries.js';
